move convert_any_value_to_json_value to traces utils
